Make both binary and maturin works well together
